requirements.txt
# Intrusion Detection using LIDAR Sensor - Essential Requirements
# Created for easy installation and testing

# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Computer Vision and Image Processing
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.5.0

# 3D Point Cloud Processing
open3d>=0.16.0
trimesh>=3.15.0

# Scientific Computing
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Handling and Visualization
pandas>=1.3.0
tqdm>=4.60.0
rich>=12.0.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.4
termcolor>=1.1.0

# MMDetection3D Core (Essential)
mmcv>=2.0.0
mmdet>=3.0.0
mmengine>=0.8.0
addict>=2.4.0

# File Handling
plyfile>=0.7.4

# Development and Testing
pytest>=6.0.0
